Compute BTRFS backing store size.
This depends on the available space and the configured size.

The default configured size (-1 MB) means that the available space on
the file system containing /var/vcap/data will be used. Since the
backing store file is sparse, this won't consume all available space
initially, but it will allow all available space to be consumed if
necessary.

[finishes #99243064]

Signed-off-by: Glyn Normington <gnormington@pivotal.io>