better idempotency/isolation of bosh-lite cgroup workaround
in working on baggageclaim we noticed the two of them both setting up this
workaround wouldn't work; they both referred to the same global workaround
directory and would only mount it if the directory didn't already exist. so if
Garden came up first it'd make the directory (which both would see) and then
mount it (which only Garden would see since it's in its own mount namespace).

the same would happen if garden_ctl would fail to set up btrfs, and then it'd
fail on the cgroup workaround the next time the ctl script ran, since it'd be
in a new mount namespace.

we'll make the same change in our own component, but figured I'd fix it here
too.